The 'Queue' command goes generally in the end of file
Moved the 'queue' or 'queue #' to the end of file, this is just a
matter or order in the file. The ordering should go further so
that the main and required information goes first, making it easier
for users to become familiar with the file structure and cluster
submission requirements.